:art: 폴더링 구조 추가, 주석 제거, 접근 제한자 반영 (PR 리뷰 반영)
